Base logic for new event addition in Admin home page.

+++ Timuon/Models/Event.cs
@@ -7,12 +7,12 @@ namespace Timuon.Models
         public string Name { get; set; }
         public DateTime Date { get; set; }
         public string Auditorium { get; set; }
-        public int Duration { get; set; }
+        public double Duration { get; set; }
         public string Audience { get; set; }
         public string Channel { get; set; }
         public string Schedule { get; set; }
         public Event(string name, DateTime date, string auditorium,
-            int duration, string audience, string channel, string schedule)
+            double duration, string audience, string channel, string schedule)
         {
             Name = name;
             Date = date;
